+++ src/main/java/pl/hycom/jira/plugins/gitlab/integration/dao/ConfigManagerDao.java
@@ -9,10 +9,10 @@ import java.util.List;
  * Created by vagrant on 5/17/16.
  */
 
-public interface ConfigManagerDao
-{
-    ConfigEntity getProjectConfig(int projectID) throws SQLException;
+public interface ConfigManagerDao {
+
+    ConfigEntity getProjectConfig(Long projectID) throws SQLException;
     public List<ConfigEntity> getAllProjectConfigs() throws SQLException;
-    ConfigEntity updateProjectConfig(int projectID,String gitlabLink,String gitlabSecret,String gitlabClientId,String gitlabProjectName) throws SQLException;
-    public void updateGitlabProjectId(int projectID, int gitlabProjectID);
+    ConfigEntity updateProjectConfig(Long projectID,String gitlabLink,String gitlabSecret,String gitlabClientId,String gitlabProjectName) throws SQLException;
+    public void updateGitlabProjectId(Long projectID, int gitlabProjectID);
 }
